+++ public/scripts/dropzoneCustom.ts
@@ -6,7 +6,7 @@ const noPyhon = 'Please only use python files!'
 Dropzone.options.zonemini = {
     createImageThumbnails: false,
     parallelUploads: 1,
-    maxFilesize: 0.05,
+    maxFilesize: 1,
     accept: (file, accept) => canAcceptFile(file, accept),
     init: dropzoneInit,
     success: (file, response) => fileGraded(file, response),
@@ -20,7 +20,7 @@ function dropzoneInit() {
 }
 
 function canAcceptFile(file, accept) {
-    if (!isPython(file)) {
+    if (!isPython(file) && projectType == "auto_code") {
         zone.removeFile(file)
         Group.showErrorNoPython(true, noPyhon)
     } else accept()
@@ -32,7 +32,7 @@ function sending(file, xhr, formData) {
 }
 
 function fileAdded(file) {
-    if (isPython(file)) Group.showErrorNoPython(false)
+    if (isPython(file) || projectType != "auto_code") Group.showErrorNoPython(false)
     else {
         zone.removeFile(file)
         Group.showErrorNoPython(true, noPyhon)
